% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strongly_connected_component.R
\name{get_strongly_connected_component}
\alias{get_strongly_connected_component}
\title{Get largest strongly connected component of a graph}
\usage{
get_strongly_connected_component(g)
}
\arguments{
\item{g}{Graph object}
}
\value{
Subgraph containing the largest strongly connected component
}
\description{
Given a weighted graph object, return the largest strongly connected component.
}
