% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_method.R
\name{power_method}
\alias{power_method}
\title{Apply 'power method' iterated fixed point algorithm to get the exp(V) vector
from Sorkin (2018).}
\usage{
power_method(g, tol = 1e-06, max_iter = 5e+05)
}
\arguments{
\item{g}{Weighted graph object}

\item{tol}{Tolerance for fixed-point algorithm}

\item{max_iter}{Maximum number of iterations for fixed-point algorithm}
}
\value{
exp(V) a vector of values (see paper for details)
}
\description{
Apply 'power method' iterated fixed point algorithm to get the exp(V) vector
from Sorkin (2018).
}
