% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprank.R
\name{rprank}
\alias{rprank}
\title{Get revealed preference ranking of firms from data on firm transitions}
\usage{
rprank(file_path, tol = 1e-06, max_iter = 5e+05)
}
\arguments{
\item{file_path}{Location of the data set}

\item{tol}{Tolerance level for convergence of the power method}

\item{max_iter}{Maximum number of iterations for the power method}
}
\value{
Data set with exponential value (expV) for each firm in the largest conneted component.
}
\description{
Given a data set containing transitions across firms,
}
